debateland/requirements.txt
Django>=3.2,<4.0
django-jsonify==0.3.0
wagtail>=4.1,<4.2
wagtail-autocomplete==0.10.0
PyICU==2.11
